% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_col_map.R
\name{add_alias}
\alias{add_alias}
\alias{add_alias,ColMap,character,character-method}
\title{add_alias}
\usage{
add_alias(x, col_name, alias)

\S4method{add_alias}{ColMap,character,character}(x, col_name, alias)
}
\arguments{
\item{x}{.}

\item{col_name}{.}

\item{alias}{.}
}
\value{
a Datafile obj
}
\description{
add_alias
}
