% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_col_map.R
\name{add_col}
\alias{add_col}
\alias{add_col,ColMap,character,character,function-method}
\title{add_col}
\usage{
add_col(
  x,
  col_name,
  col_type,
  func,
  aliases = list(col_name),
  active = TRUE,
  overwrite = FALSE
)

\S4method{add_col}{ColMap,character,character,`function`}(
  x,
  col_name,
  col_type,
  func,
  aliases = list(col_name),
  active = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{x}{ColMap object}

\item{col_name}{.}

\item{col_type}{.}

\item{func}{.}

\item{aliases}{.}

\item{active}{.}

\item{overwrite}{.}
}
\value{
ColMap object
}
\description{
add_col
}
