% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_col_map.R
\name{all_cols}
\alias{all_cols}
\alias{all_cols,ColMap-method}
\title{all_cols}
\usage{
all_cols(x)

\S4method{all_cols}{ColMap}(x)
}
\arguments{
\item{x}{a Colmap object}
}
\value{
the names of the active columns
}
\description{
all_cols
}
