% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{base_column_mapping}
\alias{base_column_mapping}
\title{my data}
\format{
An object of class \code{ColMap} of length 1.
}
\usage{
base_column_mapping
}
\description{
my data
}
\keyword{datasets}
