% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_col_map.R
\name{col_active}
\alias{col_active}
\alias{col_active,ColMap-method}
\title{col_active}
\usage{
col_active(x)

\S4method{col_active}{ColMap}(x)
}
\arguments{
\item{x}{a Colmap object}
}
\value{
logical vectors of possible columns (nameS) and whether active or not
}
\description{
col_active
}
