% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_col_map.R
\name{col_aliases}
\alias{col_aliases}
\alias{col_aliases,ColMap-method}
\title{col_aliases}
\usage{
col_aliases(x)

\S4method{col_aliases}{ColMap}(x)
}
\arguments{
\item{x}{a Colmap object}
}
\value{
a list of alias vectors
}
\description{
col_aliases
}
