% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_col_map.R
\name{col_fill}
\alias{col_fill}
\alias{col_fill,ColMap-method}
\title{col_fill}
\usage{
col_fill(x)

\S4method{col_fill}{ColMap}(x)
}
\arguments{
\item{x}{a Colmap object}
}
\value{
the fill value
}
\description{
col_fill
}
