% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_col_map.R
\name{col_map}
\alias{col_map}
\alias{col_map,ColMap-method}
\title{col_map}
\usage{
col_map(x, ...)

\S4method{col_map}{ColMap}(x, input_col_names = NULL, only.active = TRUE, ignore.case = TRUE)
}
\arguments{
\item{x}{a Colmap object}

\item{...}{.}

\item{input_col_names}{NULL,}

\item{only.active}{.}

\item{ignore.case}{.}

\item{ColMap}{}
}
\value{
the col name map
}
\description{
col_map
}
