% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_col_map.R, R/class_datafile.R
\name{col_names}
\alias{col_names}
\alias{col_names,ColMap-method}
\alias{col_names,DataFile-method}
\title{col_names}
\usage{
col_names(x)

\S4method{col_names}{ColMap}(x)

\S4method{col_names}{DataFile}(x)
}
\arguments{
\item{x}{a Colmap object}
}
\value{
names of the active columns
}
\description{
col_names
}
