% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_col_map.R
\name{col_type_funcs}
\alias{col_type_funcs}
\alias{col_type_funcs,ColMap-method}
\title{col_type_funcs}
\usage{
col_type_funcs(x)

\S4method{col_type_funcs}{ColMap}(x)
}
\arguments{
\item{x}{a Colmap object}
}
\value{
a list of functions to apply to the columns on extracting
}
\description{
col_type_funcs
}
