% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_col_map.R
\name{col_types}
\alias{col_types}
\alias{col_types,ColMap-method}
\title{col_types}
\usage{
col_types(x)

\S4method{col_types}{ColMap}(x)
}
\arguments{
\item{x}{a Colmap object}
}
\value{
the column types of the active columns
}
\description{
col_types
}
