% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_col_map.R
\docType{class}
\name{ColMap-class}
\alias{ColMap-class}
\alias{ColMap}
\title{ColMap}
\value{
a ColMap object
}
\description{
ColMap
}
\section{Slots}{

\describe{
\item{\code{active}}{logical.}

\item{\code{aliases}}{list.}

\item{\code{col_types}}{character.}

\item{\code{funcs}}{list.}

\item{\code{col_fill}}{ANY.}
}}

