% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_study.R
\name{copy_file_structure}
\alias{copy_file_structure}
\alias{copy_file_structure,list-method}
\title{copy_file_structure}
\usage{
copy_file_structure(file_structure)

\S4method{copy_file_structure}{list}(file_structure)
}
\arguments{
\item{file_structure}{.}
}
\value{
.
}
\description{
copy_file_structure
}
