% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_study.R
\name{create_data_files}
\alias{create_data_files}
\alias{create_data_files,Study,listORcharacterORmissing-method}
\title{create_data_files}
\usage{
create_data_files(object, x = NULL, append_regex = "")

\S4method{create_data_files}{Study,listORcharacterORmissing}(object, x = NULL)
}
\arguments{
\item{object}{.}

\item{x}{.}

\item{append_regex}{.}
}
\value{
.
}
\description{
create_data_files
}
