% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gwas_sumstats.R
\name{create_eafplot}
\alias{create_eafplot}
\alias{create_eafplot,data.table-method}
\title{create_eafplot}
\usage{
create_eafplot(dt, file_path = getwd(), threshold = 0.2, ...)

\S4method{create_eafplot}{data.table}(dt, file_path = getwd(), threshold = 0.2, ...)
}
\arguments{
\item{dt}{.}

\item{file_path}{.}

\item{threshold}{.}

\item{...}{.}
}
\value{
.
}
\description{
create_eafplot
}
