% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gwas_sumstats.R
\name{create_eshist}
\alias{create_eshist}
\alias{create_eshist,data.table-method}
\title{create_eshist}
\usage{
create_eshist(dt, file_path = getwd(), ...)

\S4method{create_eshist}{data.table}(dt, file_path = getwd(), ...)
}
\arguments{
\item{dt}{.}

\item{file_path}{.}

\item{...}{.}
}
\value{
.
}
\description{
create_eshist
}
