% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gwas_sumstats.R
\name{create_forestplot}
\alias{create_forestplot}
\alias{create_forestplot,data.table,character-method}
\title{create_forestplot}
\usage{
create_forestplot(dt, file_path = getwd(), ...)

\S4method{create_forestplot}{data.table,character}(dt, file_path = getwd(), ...)
}
\arguments{
\item{dt}{.}

\item{file_path}{.}

\item{...}{.}
}
\value{
.
}
\description{
create_forestplot
}
