% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gwas_sumstats.R
\name{create_impqualhist}
\alias{create_impqualhist}
\alias{create_impqualhist,data.table-method}
\title{create_impqualhist}
\usage{
create_impqualhist(dt, file_path = getwd(), ...)

\S4method{create_impqualhist}{data.table}(dt, file_path = getwd(), ...)
}
\arguments{
\item{dt}{.}

\item{file_path}{.}

\item{...}{.}
}
\value{
.
}
\description{
create_impqualhist
}
