% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gwas_sumstats.R
\name{create_locuszoom}
\alias{create_locuszoom}
\alias{create_locuszoom,data.table,character-method}
\title{create_locuszoom}
\usage{
create_locuszoom(dt, snp, file_path = getwd(), ...)

\S4method{create_locuszoom}{data.table,character}(dt, snp, file_path = getwd(), ...)
}
\arguments{
\item{dt}{.}

\item{snp}{.}

\item{file_path}{.}

\item{...}{.}
}
\value{
.
}
\description{
create_locuszoom
}
