% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gwas_sumstats.R
\name{create_manhattan}
\alias{create_manhattan}
\alias{create_manhattan,data.table-method}
\title{create_manhattan}
\usage{
create_manhattan(
  dt,
  file_path = getwd(),
  highlight = NULL,
  annotate = 5e-08,
  ...
)

\S4method{create_manhattan}{data.table}(
  dt,
  file_path = getwd(),
  highlight = NULL,
  annotate = 5e-08,
  ...
)
}
\arguments{
\item{dt}{.}

\item{file_path}{.}

\item{highlight}{.}

\item{...}{.}
}
\value{
.
}
\description{
create_manhattan
}
