% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gwas_sumstats.R
\name{create_pvalhist}
\alias{create_pvalhist}
\alias{create_pvalhist,data.table-method}
\title{create_pvalhist}
\usage{
create_pvalhist(dt, file_path = getwd(), ...)

\S4method{create_pvalhist}{data.table}(dt, file_path = getwd(), ...)
}
\arguments{
\item{dt}{.}

\item{file_path}{.}

\item{...}{.}
}
\value{
.
}
\description{
create_pvalhist
}
