% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gwas_sumstats.R
\name{create_pz}
\alias{create_pz}
\alias{create_pz,data.table-method}
\title{create_pz}
\usage{
create_pz(dt, file_path = getwd(), ...)

\S4method{create_pz}{data.table}(dt, file_path = getwd(), ...)
}
\arguments{
\item{dt}{.}

\item{file_path}{.}

\item{...}{.}
}
\value{
.
}
\description{
create_pz
}
