% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_study.R
\name{create_qc_data_files}
\alias{create_qc_data_files}
\alias{create_qc_data_files,Study-method}
\title{create_qc_data_files}
\usage{
create_qc_data_files(object, ...)

\S4method{create_qc_data_files}{Study}(object, input_list = NULL, parent_names = character(0))
}
\arguments{
\item{object}{.}

\item{...}{.}

\item{input_list}{.}

\item{parent_names}{.}
}
\value{
.
}
\description{
create_qc_data_files
}
