% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gwas_sumstats.R
\name{create_qq}
\alias{create_qq}
\alias{create_qq,data.table-method}
\title{create_qq}
\usage{
create_qq(dt, file_path = getwd(), ...)

\S4method{create_qq}{data.table}(dt, file_path = getwd(), ...)
}
\arguments{
\item{dt}{.}

\item{file_path}{.}

\item{...}{.}
}
\value{
.
}
\description{
create_qq
}
