% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_studycorpus.R
\name{create_results_list}
\alias{create_results_list}
\alias{create_results_list,StudyCorpus,character,character-method}
\title{create_results_list}
\usage{
create_results_list(corpus, study_type, results_dir)

\S4method{create_results_list}{StudyCorpus,character,character}(corpus, study_type, results_dir)
}
\arguments{
\item{study_type}{.}

\item{results_dir}{description}

\item{object}{.}
}
\value{
.
}
\description{
create_results_list
}
