% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_studycorpus.R
\name{create_study_list}
\alias{create_study_list}
\alias{create_study_list,StudyCorpus,character-method}
\title{create_study_list}
\usage{
create_study_list(object, study_type)

\S4method{create_study_list}{StudyCorpus,character}(object, study_type)
}
\arguments{
\item{object}{.}

\item{study_type}{.}
}
\value{
.
}
\description{
create_study_list
}
