% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_datafile.R
\name{data_exists}
\alias{data_exists}
\alias{data_exists,DataFile-method}
\title{data_exists}
\usage{
data_exists(x)

\S4method{data_exists}{DataFile}(x)
}
\arguments{
\item{x}{.}
}
\value{
a Datafile obj
}
\description{
data_exists
}
