% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gwas_sumstats.R
\name{easy_qc}
\alias{easy_qc}
\alias{easy_qc,GWASsumstats-method}
\title{Title}
\usage{
easy_qc(object, ...)

\S4method{easy_qc}{GWASsumstats}(object, ...)
}
\arguments{
\item{object}{GWASsumstats oobject}

\item{...}{data_file keys}
}
\value{
GWASsumstats oobject
}
\description{
Title
}
