% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_datafile.R
\name{extract}
\alias{extract}
\alias{extract,DataFile,missing-method}
\alias{extract,DataFile,character-method}
\alias{extract,list,character-method}
\title{extract}
\usage{
extract(x, merge_col)

\S4method{extract}{DataFile,missing}(x)

\S4method{extract}{DataFile,character}(x, merge_col = "JOIN")

\S4method{extract}{list,character}(x, merge_col = "JOIN")
}
\arguments{
\item{x}{.}

\item{merge_col}{description}

\item{...}{.}
}
\value{
a Datafile obj
}
\description{
extract
}
