% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_datafile.R
\name{file_name}
\alias{file_name}
\alias{file_name,DataFile-method}
\title{file_name}
\usage{
file_name(x)

\S4method{file_name}{DataFile}(x)
}
\arguments{
\item{x}{.}
}
\value{
the filename
}
\description{
file_name
}
