% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_datafile.R
\name{get_data}
\alias{get_data}
\alias{get_data,DataFile-method}
\title{get_data}
\usage{
get_data(x)

\S4method{get_data}{DataFile}(x)
}
\arguments{
\item{x}{a DataFile object}
}
\value{
a data.table
}
\description{
get_data
}
