% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_study.R
\name{get_data_keys}
\alias{get_data_keys}
\alias{get_data_keys,Study-method}
\alias{get_data_keys,list-method}
\alias{get_data_keys,DataFile-method}
\alias{get_data_keys,character-method}
\title{get_data_keys}
\usage{
get_data_keys(input, ...)

\S4method{get_data_keys}{Study}(input, prefix = NULL)

\S4method{get_data_keys}{list}(input, prefix = NULL)

\S4method{get_data_keys}{DataFile}(input, prefix = NULL)

\S4method{get_data_keys}{character}(input, ...)
}
\arguments{
\item{input}{.}

\item{...}{.}

\item{prefix}{description}
}
\value{
.
}
\description{
get_data_keys
}
