% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gwas_sumstats.R
\docType{class}
\name{GWASsumstats-class}
\alias{GWASsumstats-class}
\alias{GWASsumstats}
\title{Title}
\value{
object
}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{parameter}}{numeric.}
}}

