% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_study.R
\name{keys_valid}
\alias{keys_valid}
\alias{keys_valid,Study,list-method}
\alias{keys_valid,Study,character-method}
\title{keys_valid}
\usage{
keys_valid(object, keys, ...)

\S4method{keys_valid}{Study,list}(object, keys)

\S4method{keys_valid}{Study,character}(object, keys, ...)
}
\arguments{
\item{object}{.}

\item{keys}{.}

\item{...}{.}
}
\value{
.
}
\description{
keys_valid
}
