% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_datafile.R
\name{mapping<-}
\alias{mapping<-}
\alias{mapping<-,list-method}
\alias{mapping<-,DataFile-method}
\title{mapping}
\usage{
mapping(x) <- value

\S4method{mapping}{list}(x) <- value

\S4method{mapping}{DataFile}(x) <- value
}
\arguments{
\item{x}{.}

\item{value}{.}
}
\value{
.
}
\description{
mapping
}
