% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_datafile.R
\name{mapping}
\alias{mapping}
\alias{mapping,DataFile-method}
\alias{mapping,list-method}
\title{mapping}
\usage{
mapping(x)

\S4method{mapping}{DataFile}(x)

\S4method{mapping}{list}(x)
}
\arguments{
\item{x}{.}
}
\value{
.map
}
\description{
mapping
}
