% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_datafile.R
\name{path<-}
\alias{path<-}
\alias{path<-,DataFile-method}
\title{set path}
\usage{
path(x) <- value

\S4method{path}{DataFile}(x) <- value
}
\arguments{
\item{x}{DataFile object}

\item{value}{the path}
}
\value{
DataFile object
}
\description{
set path
}
