% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_study.R
\name{post_qc_data_dir}
\alias{post_qc_data_dir}
\alias{post_qc_data_dir,Study-method}
\title{post_qc_data_dir}
\usage{
post_qc_data_dir(x)

\S4method{post_qc_data_dir}{Study}(x)
}
\arguments{
\item{x}{.}
}
\value{
.
}
\description{
post_qc_data_dir
}
