% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_study.R
\name{pre_qc_data_dir}
\alias{pre_qc_data_dir}
\alias{pre_qc_data_dir,Study-method}
\title{pre_qc_data_dir}
\usage{
pre_qc_data_dir(x)

\S4method{pre_qc_data_dir}{Study}(x)
}
\arguments{
\item{x}{.}
}
\value{
.
}
\description{
pre_qc_data_dir
}
