% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gwas_sumstats.R
\name{qc_filter_summary}
\alias{qc_filter_summary}
\alias{qc_filter_summary,GWASsumstats-method}
\title{qc_filter_summary}
\usage{
qc_filter_summary(object, ...)

\S4method{qc_filter_summary}{GWASsumstats}(object, ...)
}
\arguments{
\item{object}{.}

\item{...}{.}
}
\value{
.
}
\description{
qc_filter_summary
}
