% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_col_map.R
\name{remove_alias}
\alias{remove_alias}
\alias{remove_alias,ColMap,character,character-method}
\title{remove_alias}
\usage{
remove_alias(x, col_name, alias)

\S4method{remove_alias}{ColMap,character,character}(x, col_name, alias)
}
\arguments{
\item{x}{.}

\item{col_name}{.}

\item{alias}{.}
}
\value{
a Datafile obj
}
\description{
remove_alias
}
