% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_col_map.R
\name{remove_col}
\alias{remove_col}
\alias{remove_col,ColMap,character-method}
\title{remove_col}
\usage{
remove_col(x, col_name)

\S4method{remove_col}{ColMap,character}(x, col_name)
}
\arguments{
\item{x}{ColMap object}

\item{col_name}{.}
}
\value{
a Colmap object
}
\description{
remove_col
}
