% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_studycorpus.R
\name{results}
\alias{results}
\alias{results,StudyCorpus-method}
\title{results}
\usage{
results(object, ...)

\S4method{results}{StudyCorpus}(object, index = NULL)
}
\arguments{
\item{object}{.}

\item{...}{.}

\item{index}{.}
}
\value{
.
}
\description{
results
}
