% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_studycorpus.R
\name{run_filter_summary_plots}
\alias{run_filter_summary_plots}
\alias{run_filter_summary_plots,StudyCorpus,character-method}
\title{run_filter_summary_plots}
\usage{
run_filter_summary_plots(object, output_dir, parallel_cores = NA_integer_)

\S4method{run_filter_summary_plots}{StudyCorpus,character}(object, output_dir, parallel_cores = NA_integer_)
}
\arguments{
\item{object}{.}

\item{output_dir}{.}
}
\value{
True
}
\description{
run_filter_summary_plots
}
