% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_studycorpus.R
\name{run_gwama}
\alias{run_gwama}
\alias{run_gwama,StudyCorpus,character-method}
\title{run_gwama}
\usage{
run_gwama(corpus, output_dir, ...)

\S4method{run_gwama}{StudyCorpus,character}(corpus, output_dir, ...)
}
\arguments{
\item{corpus}{.}

\item{output_dir}{.}

\item{...}{.}
}
\value{
.
}
\description{
run_gwama
}
