% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_studycorpus.R
\name{run_meta_plots}
\alias{run_meta_plots}
\alias{run_meta_plots,StudyCorpus,character-method}
\title{run_meta_plots}
\usage{
run_meta_plots(corpus, output_dir, index = NULL)

\S4method{run_meta_plots}{StudyCorpus,character}(corpus, output_dir, index = NULL)
}
\arguments{
\item{output_dir}{.}

\item{index}{.}

\item{object}{.}

\item{...}{.}

\item{parallel_cores}{.}
}
\value{
True
}
\description{
run_meta_plots
}
