% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gwas_sumstats.R
\name{run_qc}
\alias{run_qc}
\alias{run_qc,GWASsumstats-method}
\title{run_qc}
\usage{
run_qc(object, ...)

\S4method{run_qc}{GWASsumstats}(object, ...)
}
\arguments{
\item{object}{.}

\item{...}{.}
}
\value{
.
}
\description{
run_qc
}
