% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gwas_sumstats.R
\name{run_qc_plots}
\alias{run_qc_plots}
\title{Title}
\usage{
run_qc_plots(object, output_dir, ...)
}
\arguments{
\item{object}{.}

\item{output_dir}{.}

\item{...}{}
}
\value{
side effects, graphs
}
\description{
Title
}
