% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_datafile.R
\name{set_data<-}
\alias{set_data<-}
\alias{set_data<-,DataFile-method}
\title{set_data}
\usage{
set_data(x) <- value

\S4method{set_data}{DataFile}(x) <- value
}
\arguments{
\item{x}{.}

\item{value}{.}
}
\value{
.
}
\description{
set_data
}
