% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_studycorpus.R
\name{studies<-}
\alias{studies<-}
\alias{studies<-,StudyCorpus-method}
\title{studies-set}
\usage{
studies(x, ...) <- value

\S4method{studies}{StudyCorpus}(x, index = NULL) <- value
}
\arguments{
\item{x}{.}

\item{value}{.}

\item{...}{. .}

\item{index}{.}
}
\value{
.
}
\description{
studies-set
}
