% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_studycorpus.R
\name{studies}
\alias{studies}
\alias{studies,StudyCorpus-method}
\title{studies}
\usage{
studies(object, ...)

\S4method{studies}{StudyCorpus}(object, index = NULL)
}
\arguments{
\item{object}{.}

\item{...}{.}

\item{index}{.}
}
\value{
.
}
\description{
studies
}
