% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_study.R
\docType{class}
\name{Study-class}
\alias{Study-class}
\alias{Study}
\title{Title}
\value{
object
}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{dir}}{character.}

\item{\code{pmid}}{integer.}

\item{\code{file_structure}}{list.}

\item{\code{col_names}}{character.}

\item{\code{col_types}}{list.}

\item{\code{data_files}}{list.}
}}

