% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_studycorpus.R
\docType{class}
\name{StudyCorpus-class}
\alias{StudyCorpus-class}
\alias{StudyCorpus}
\title{Title}
\value{
StudyCorpus object
}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{corpus_dir}}{character.}

\item{\code{study_type}}{character.}

\item{\code{studies}}{list.}

\item{\code{results}}{.}

\item{\code{meta_quantitative}}{.}

\item{\code{meta_indel_alleles}}{.}
}}

