% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_studycorpus.R
\name{[,StudyCorpus,ANY,ANY,ANY-method}
\alias{[,StudyCorpus,ANY,ANY,ANY-method}
\title{Title}
\usage{
\S4method{[}{StudyCorpus,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{x}{..}

\item{i}{.}
}
\value{
.
}
\description{
Title
}
