% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_datafile.R
\name{write_file}
\alias{write_file}
\alias{write_file,DataFile,character-method}
\alias{free,list-method}
\alias{free,DataFile-method}
\title{write_file}
\usage{
write_file(x, file_path, overwrite = FALSE, append = FALSE, sep = "\\t", ...)

\S4method{write_file}{DataFile,character}(x, file_path, overwrite = FALSE, append = FALSE, sep = "\\t", ...)

\S4method{free}{list}(x)

\S4method{free}{DataFile}(x)
}
\arguments{
\item{x}{.}

\item{file_path}{.}

\item{overwrite}{.}

\item{append}{.}

\item{sep}{.}

\item{...}{.}
}
\value{
write_file
}
\description{
write_file
}
