% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_col_map.R
\name{col_names}
\alias{col_names}
\alias{col_names,ColMap-method}
\title{columns}
\usage{
col_names(x)

\S4method{col_names}{ColMap}(x)
}
\arguments{
\item{x}{ColMap object}
}
\value{
active column names
}
\description{
columns
}
