% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gwas_sumstats.R
\name{create_heatmap}
\alias{create_heatmap}
\alias{create_heatmap,data.table-method}
\title{create_heatmap}
\usage{
create_heatmap(
  dt,
  x,
  y,
  z,
  fx = NULL,
  fy = NULL,
  x_lab = "x",
  y_lab = "y",
  fill_lab = "fill",
  fill_lims = c(0, 1),
  h = 1200,
  w = 1200,
  file_path = getwd(),
  ...
)

\S4method{create_heatmap}{data.table}(
  dt,
  x,
  y,
  z,
  fx = NULL,
  fy = NULL,
  x_lab = "x",
  y_lab = "y",
  fill_lab = "fill",
  fill_lims = c(0, 1),
  h = 1200,
  w = 1200,
  file_path = getwd(),
  ...
)
}
\arguments{
\item{dt}{.}

\item{file_path}{.}

\item{...}{.}
}
\value{
.
}
\description{
create_heatmap
}
