% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_datafile.R
\docType{class}
\name{DataFile-class}
\alias{DataFile-class}
\alias{DataFile}
\title{DataFile class}
\value{
a DataFile object
}
\description{
DataFile class
}
\section{Slots}{

\describe{
\item{\code{path}}{character.}

\item{\code{data}}{data.table.}

\item{\code{mapping}}{ColMap}
}}

