% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gwas_sumstats.R
\docType{class}
\name{GWASsumstats-class}
\alias{GWASsumstats-class}
\alias{GWASsumstats}
\title{Title}
\value{
object
}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{qc_freq_threshold}}{.}

\item{\code{qc_low_events_n}}{,}

\item{\code{qc_freq_low_events_thresh}}{.}

\item{\code{qc_freq_ambig_thresh}}{.}

\item{\code{qc_qual_threshold}}{.}

\item{\code{qc_call_rate_threshold}}{.}

\item{\code{qc_freq_diff_threshold}}{.}

\item{\code{ref_path}}{.}

\item{\code{ref_data_file}}{.}
}}

