% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_studycorpus.R
\name{run_meta_plots}
\alias{run_meta_plots}
\alias{run_meta_plots,StudyCorpus,character-method}
\title{run_meta_plots}
\usage{
run_meta_plots(object, output_dir, index = NULL, parallel_cores = NA_integer_)

\S4method{run_meta_plots}{StudyCorpus,character}(object, output_dir, index = NULL, parallel_cores = NA_integer_)
}
\arguments{
\item{object}{.}

\item{output_dir}{.}

\item{index}{.}

\item{parallel_cores}{.}

\item{...}{.}
}
\value{
True
}
\description{
run_meta_plots
}
