% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gwas_sumstats.R, R/class_studycorpus.R
\name{run_qc}
\alias{run_qc}
\alias{run_qc,GWASsumstats-method}
\alias{run_qc,StudyCorpus-method}
\title{run_qc}
\usage{
run_qc(object, ...)

\S4method{run_qc}{GWASsumstats}(object, ...)

\S4method{run_qc}{StudyCorpus}(object, ..., index = NULL, parallel_cores = NA_integer_)
}
\arguments{
\item{object}{.}

\item{...}{.}

\item{index}{.}

\item{parallel_cores}{.}
}
\value{
.
}
\description{
run_qc
}
