% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gwas_sumstats.R, R/class_studycorpus.R
\name{run_qc_plots}
\alias{run_qc_plots}
\alias{run_qc_plots,GWASsumstats,character-method}
\alias{run_qc_plots,StudyCorpus,character-method}
\title{run_qc_plots}
\usage{
run_qc_plots(object, output_dir, ...)

\S4method{run_qc_plots}{GWASsumstats,character}(object, output_dir, ...)

\S4method{run_qc_plots}{StudyCorpus,character}(
  object,
  output_dir,
  ...,
  index = NULL,
  parallel_cores = NA_integer_
)
}
\arguments{
\item{object}{.}

\item{output_dir}{.}

\item{...}{.}

\item{index}{.}

\item{parallel_cores}{.}
}
\value{
.
}
\description{
run_qc_plots
}
