% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_studycorpus.R
\name{studies}
\alias{studies}
\alias{studies,StudyCorpus-method}
\alias{studies<-,StudyCorpus-method}
\title{studies}
\usage{
studies(object, ...)

\S4method{studies}{StudyCorpus}(object, index = NULL)

\S4method{studies}{StudyCorpus}(x, index = NULL) <- value
}
\arguments{
\item{object}{.}
}
\value{
.
}
\description{
studies
}
