% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_study.R
\docType{class}
\name{Study-class}
\alias{Study-class}
\alias{Study}
\title{Study virtual class}
\value{
object
}
\description{
Study virtual class
}
\section{Slots}{

\describe{
\item{\code{dir}}{character.}

\item{\code{file_structure}}{list.}

\item{\code{data_files}}{list.}

\item{\code{pre_qc_dir}}{.}

\item{\code{post_qc_dir}}{.}

\item{\code{mapping}}{.}

\item{\code{qc_data_files}}{.}
}}

